client/requirements.txt
fastapi
uvicorn[standard]
pymysql
pandas
python-dotenv
redis
httpx
pydantic
